[ESWE-858] Matching candidate job details distance calculation (#113)
This PR aims to introduce the distance calculation between two locations
identified by a postal code. We must first obtain the coordinates
defining each postcode to do this calculation.

### Postcodes to coordinates

Each postcode has a pair of coordinates XY associated that defines its
reference centre point. To obtain it, we use a third-party service, in
our case [OS Places
API](https://osdatahub.os.uk/docs/places/technicalSpecification)
The keys for use on our environments have been requested via the
[#ask-operations-engineering](https://mojdt.slack.com/archives/C01BUKJSZD4/p1728314239141029)
Slack channel.

The keys are stored on Kubernetes as secrets.

### Intermediate storage

One of the essential architectural decisions taken here is to store the
postcodes and their coordinates in our database as a primitive cache.
The main reasons were:
- If the third-party service is down, we can still provide calculations.
- If the postcode is already stored on the database, the request won't
be done, saving time and resources.
- To manage calculations at a pagination level, we can improve its
efficiency by moving that logic to the database query itself.

### Cache trade-offs:
Considering a simple cache implies some maintenance and decisions. For
example, Royal Mail generates, modifies, deletes, and reuses postal
codes monthly. But they get to the services every three months. We must
assume that those coordinates we store have an expiration date and need
to be refreshed occasionally. This is out of the scope of this PR, but
that work must be planned to make this solution last without introducing
errors. For now, the creation and update dates are stored for this
purpose in the future.

### When we store the coordinates
Every time a Job is created or updated, the coordinates related to its
postcode are retrieved and stored in the database. That makes updates an
excellent tool to recover missing coordinates if something happens on
creation.

<img width="1255" alt="image"
src="https://github.com/user-attachments/assets/ffdb16e2-8ae0-44b3-9449-943e97e8cf57">

Whenever a search for a candidate matching list or detail is invoked,
the release area postcode coordinates are retrieved and stored, being
ready to calculate the distance.

<img width="1230" alt="image"
src="https://github.com/user-attachments/assets/1e52f577-df25-41ff-8d3d-59121137e15f">

+++ src/main/kotlin/uk/gov/justice/digital/hmpps/jobsboard/api/jobs/domain/Postcode.kt
@@ -16,10 +16,10 @@ data class Postcode(
   var code: String,
 
   @Column(name = "x_coordinate", nullable = true)
-  var xCoordinate: Float?,
+  var xCoordinate: Double?,
 
   @Column(name = "y_coordinate", nullable = true)
-  var yCoordinate: Float?,
+  var yCoordinate: Double?,
 ) : Auditable() {
   override fun toString(): String = """
     Postcode(
