+++ src/main/kotlin/uk/gov/justice/digital/hmpps/jobsboard/api/jobs/infrastructure/OsPlacesApiDPA.kt
@@ -7,8 +7,8 @@ data class OsPlacesApiDPA(
   val postcode: String,
 
   @JsonProperty("X_COORDINATE")
-  val xCoordinate: Float?,
+  val xCoordinate: Double?,
 
   @JsonProperty("Y_COORDINATE")
-  val yCoordinate: Float?,
+  val yCoordinate: Double?,
 )
